simplify, improve efficiency
